+++ q2_types/__init__.py
@@ -19,7 +19,6 @@ importlib.import_module('q2_types.sample_data')
 importlib.import_module('q2_types.metadata')
 importlib.import_module('q2_types.multiplexed_sequences')
 importlib.import_module('q2_types.kraken2')
-importlib.import_module('q2_types.genome_data')
 importlib.import_module('q2_types.kaiju')
 importlib.import_module('q2_types.reference_db')
 importlib.import_module('q2_types.profile_hmms')
