+++ q2_types/plugin_setup.py
@@ -34,5 +34,6 @@ importlib.import_module('q2_types.feature_data._deferred_setup')
 importlib.import_module('q2_types.feature_data_mag._deferred_setup')
 importlib.import_module('q2_types.feature_map._deferred_setup')
 importlib.import_module('q2_types.feature_table._deferred_setup')
+importlib.import_module('q2_types.genome_data._deferred_setup')
 
 importlib.import_module('q2_types.per_sample_sequences._deferred_setup')
